+++ packages/config-helpers/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.4.1](https://github.com/eslint/rewrite/compare/config-helpers-v0.4.0...config-helpers-v0.4.1) (2025-10-17)
+
+
+### Bug Fixes
+
+* add validation for `plugins`  in isLegacyConfig ([#292](https://github.com/eslint/rewrite/issues/292)) ([74f9427](https://github.com/eslint/rewrite/commit/74f9427b47de313582793ab6fc4c723f1526fdc0))
+* improve type support for isolated dependencies in pnpm ([#289](https://github.com/eslint/rewrite/issues/289)) ([f8df139](https://github.com/eslint/rewrite/commit/f8df139631694431ecfc651e656932e283d4d14f))
+* use flat config when eslintrc config does not exist ([#288](https://github.com/eslint/rewrite/issues/288)) ([ddc8577](https://github.com/eslint/rewrite/commit/ddc857781bacab1cdd7c540e599d3ed968607a09))
+
 ## [0.4.0](https://github.com/eslint/rewrite/compare/config-helpers-v0.3.1...config-helpers-v0.4.0) (2025-09-16)